Add sig count check to validatation
